--- src/BoltOn.Caching/BoltOn.Caching.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <Version>0.1.0</Version>
-    <Copyright>BoltOn</Copyright>
-    <Authors>Gokulnath Murugesan</Authors>
-    <Owners>Gokulnath Murugesan</Owners>
-    <PackageProjectUrl>https://github.com/gokulm/BoltOn</PackageProjectUrl>
-    <PackageTags>C#, .NET</PackageTags>
-    <Title>BoltOn.Caching</Title>
-    <PackageId>BoltOn.Caching</PackageId>
-    <Description>This is an add-on to BoltOn NuGet package to support Caching</Description>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\BoltOn\BoltOn.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="3.1.3" />
-    <PackageReference Include="System.Text.Json" Version="4.7.1" />
-  </ItemGroup>
-</Project>
