--- src/BoltOn.Caching/BoltOnCacheSerializer.cs
@@ -1,26 +1,0 @@
-﻿using System.Text.Json;
-
-namespace BoltOn.Caching
-{
-    public interface IBoltOnCacheSerializer
-    {
-        byte[] ToByteArray(object obj);
-        T FromByteArray<T>(byte[] byteArray) where T : class;
-    }
-
-    public class BoltOnCacheSerializer : IBoltOnCacheSerializer
-    {
-        public T FromByteArray<T>(byte[] byteArray) where T : class
-        {
-            if (byteArray == default || byteArray.Length == 0)
-                return default;
-
-            return JsonSerializer.Deserialize<T>(byteArray);
-        }
-
-        public byte[] ToByteArray(object obj)
-        {
-            return JsonSerializer.SerializeToUtf8Bytes(obj);
-        }
-    }
-}
