--- src/BoltOn.Caching/Extensions.cs
@@ -1,16 +1,0 @@
-using System.Reflection;
-using BoltOn.Bootstrapping;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace BoltOn.Caching
-{
-	public static class Extensions
-	{
-		public static BoltOnOptions BoltOnCaching(this BoltOnOptions boltOnOptions)
-		{
-			boltOnOptions.BoltOnAssemblies(Assembly.GetExecutingAssembly());
-			boltOnOptions.ServiceCollection.AddTransient<IBoltOnCache, BoltOnCache>();
-			return boltOnOptions;
-		}
-	}
-}
