+++ tests/BoltOn.Tests/BoltOn.Tests.csproj
@@ -42,14 +42,14 @@
         <Folder Include="Data\CosmosDb\" />
         <Folder Include="Cqrs\Fakes\" />
         <Folder Include="Bootstrapping\Fakes\" />
-        <Folder Include="Caching\" />
-        <Folder Include="Caching\Fakes\" />
+        <Folder Include="Cache\" />
+        <Folder Include="Cache\Fakes\" />
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\..\src\BoltOn\BoltOn.csproj" />
         <ProjectReference Include="..\..\src\BoltOn.Data.EF\BoltOn.Data.EF.csproj" />
         <ProjectReference Include="..\..\src\BoltOn.Bus.MassTransit\BoltOn.Bus.MassTransit.csproj" />
         <ProjectReference Include="..\..\src\BoltOn.Data.CosmosDb\BoltOn.Data.CosmosDb.csproj" />
-        <ProjectReference Include="..\..\src\BoltOn.Caching\BoltOn.Caching.csproj" />
+        <ProjectReference Include="..\..\src\BoltOn.Cache\BoltOn.Cache.csproj" />
     </ItemGroup>
 </Project>
