+++ tests/BoltOn.Tests/Cache/BoltOnCacheTests.cs
@@ -1,15 +1,15 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using BoltOn.Caching;
+using BoltOn.Cache;
 using BoltOn.Logging;
-using BoltOn.Tests.Caching.Fakes;
+using BoltOn.Tests.Cache.Fakes;
 using Microsoft.Extensions.Caching.Distributed;
 using Moq;
 using Moq.AutoMock;
 using Xunit;
 
-namespace BoltOn.Tests.Caching
+namespace BoltOn.Tests.Cache
 {
 	public class BoltOnCacheTests
 	{
