--- tests/BoltOn.Tests/Caching/Fakes/CacheTestHelper.cs
@@ -1,10 +1,0 @@
-using System.IO;
-using System.Text.Json;
-using System.Runtime.Serialization.Formatters.Binary;
-
-namespace BoltOn.Tests.Caching.Fakes
-{
-	public class CacheTestHelper
-	{
-	}
-}
